fix: Change fan RPM type from f32 to f64 for precision
- Updated the `front_fan_rpm` and `back_fan_rpm` fields in `LiveValuesEvent` struct from f32 to f64.
- Adjusted the `FanState` struct to use f64 for the `rpm` field.
- Modified the `SetFrontFan` and `SetBackFan` mutation variants to expect f64 values instead of f32.

+++ machines/src/aquapath1/api.rs
@@ -20,8 +20,8 @@ pub struct LiveValuesEvent {
     pub back_temperature: f64,
     pub front_temp_reservoir: f64,
     pub back_temp_reservoir: f64,
-    pub front_fan_rpm: f32,
-    pub back_fan_rpm: f32,
+    pub front_fan_rpm: f64,
+    pub back_fan_rpm: f64,
 }
 
 impl LiveValuesEvent {
@@ -75,7 +75,7 @@ pub struct FlowState {
 
 #[derive(Serialize, Debug, Clone)]
 pub struct FanState {
-    pub rpm: f32,
+    pub rpm: f64,
 }
 
 #[derive(Serialize, Debug, Clone)]
@@ -100,8 +100,8 @@ enum Mutation {
     SetFrontFlow(bool),
     SetBackFlow(bool),
 
-    SetBackFan(f32),
-    SetFrontFan(f32),
+    SetBackFan(f64),
+    SetFrontFan(f64),
 }
 
 #[derive(Debug, Clone)]
